ASM. - Deleted my random number generator. - Changed MOVQ to MOVSD - Begining of implementing _Init

+++ Perlin.PureC/Perlin.PureC.c
@@ -32,7 +32,7 @@ PERLINPUREC_API void Init(size_t width, size_t height)
 		p[i] = i;
 
 		for (j = 0; j < 2; j++)
-			g2[i][j] = (float)((rand() % (B + B)) - B) / B;
+			g2[i][j] = (double)((rand() % (B + B)) - B) / B;
 
 		normalize(g2[i]);
 	}
